refactor(ui): split out the hook isolator component
