+++ invokeai/frontend/web/src/app/components/App.tsx
@@ -1,18 +1,12 @@
-import { Box, useGlobalModifiersInit } from '@invoke-ai/ui-library';
+import { Box } from '@invoke-ai/ui-library';
 import { useStore } from '@nanostores/react';
+import { GlobalHookIsolator } from 'app/components/GlobalHookIsolator';
 import { GlobalImageHotkeys } from 'app/components/GlobalImageHotkeys';
 import type { StudioInitAction } from 'app/hooks/useStudioInitAction';
-import { $didStudioInit, useStudioInitAction } from 'app/hooks/useStudioInitAction';
-import { useSyncQueueStatus } from 'app/hooks/useSyncQueueStatus';
-import { useLogger } from 'app/logging/useLogger';
-import { useSyncLoggingConfig } from 'app/logging/useSyncLoggingConfig';
-import { appStarted } from 'app/store/middleware/listenerMiddleware/listeners/appStarted';
-import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
+import { $didStudioInit } from 'app/hooks/useStudioInitAction';
 import type { PartialAppConfig } from 'app/types/invokeai';
 import Loading from 'common/components/Loading/Loading';
-import { useFocusRegionWatcher } from 'common/hooks/focus';
 import { useClearStorage } from 'common/hooks/useClearStorage';
-import { useGlobalHotkeys } from 'common/hooks/useGlobalHotkeys';
 import ChangeBoardModal from 'features/changeBoardModal/components/ChangeBoardModal';
 import { CanvasPasteModal } from 'features/controlLayers/components/CanvasPasteModal';
 import {
@@ -25,30 +19,22 @@ import { FullscreenDropzone } from 'features/dnd/FullscreenDropzone';
 import { DynamicPromptsModal } from 'features/dynamicPrompts/components/DynamicPromptsPreviewModal';
 import DeleteBoardModal from 'features/gallery/components/Boards/DeleteBoardModal';
 import { ImageContextMenu } from 'features/gallery/components/ImageContextMenu/ImageContextMenu';
-import { useStarterModelsToast } from 'features/modelManagerV2/hooks/useStarterModelsToast';
 import { ShareWorkflowModal } from 'features/nodes/components/sidePanel/workflow/WorkflowLibrary/ShareWorkflowModal';
 import { WorkflowLibraryModal } from 'features/nodes/components/sidePanel/workflow/WorkflowLibrary/WorkflowLibraryModal';
 import { CancelAllExceptCurrentQueueItemConfirmationAlertDialog } from 'features/queue/components/CancelAllExceptCurrentQueueItemConfirmationAlertDialog';
 import { ClearQueueConfirmationsAlertDialog } from 'features/queue/components/ClearQueueConfirmationAlertDialog';
-import { useReadinessWatcher } from 'features/queue/store/readiness';
 import { DeleteStylePresetDialog } from 'features/stylePresets/components/DeleteStylePresetDialog';
 import { StylePresetModal } from 'features/stylePresets/components/StylePresetForm/StylePresetModal';
 import RefreshAfterResetModal from 'features/system/components/SettingsModal/RefreshAfterResetModal';
 import { VideosModal } from 'features/system/components/VideosModal/VideosModal';
-import { configChanged } from 'features/system/store/configSlice';
-import { selectLanguage } from 'features/system/store/systemSelectors';
 import { AppContent } from 'features/ui/components/AppContent';
 import { DeleteWorkflowDialog } from 'features/workflowLibrary/components/DeleteLibraryWorkflowConfirmationAlertDialog';
 import { LoadWorkflowConfirmationAlertDialog } from 'features/workflowLibrary/components/LoadWorkflowConfirmationAlertDialog';
 import { LoadWorkflowFromGraphModal } from 'features/workflowLibrary/components/LoadWorkflowFromGraphModal/LoadWorkflowFromGraphModal';
 import { NewWorkflowConfirmationAlertDialog } from 'features/workflowLibrary/components/NewWorkflowConfirmationAlertDialog';
 import { SaveWorkflowAsDialog } from 'features/workflowLibrary/components/SaveWorkflowAsDialog';
-import i18n from 'i18n';
-import { size } from 'lodash-es';
-import { memo, useCallback, useEffect } from 'react';
+import { memo, useCallback } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
-import { useGetOpenAPISchemaQuery } from 'services/api/endpoints/appInfo';
-import { useSocketIO } from 'services/events/useSocketIO';
 
 import AppErrorBoundaryFallback from './AppErrorBoundaryFallback';
 const DEFAULT_CONFIG = {};
@@ -74,53 +60,13 @@ const App = ({ config = DEFAULT_CONFIG, studioInitAction }: Props) => {
         <AppContent />
         {!didStudioInit && <Loading />}
       </Box>
-      <HookIsolator config={config} studioInitAction={studioInitAction} />
+      <GlobalHookIsolator config={config} studioInitAction={studioInitAction} />
       <ModalIsolator />
     </ErrorBoundary>
   );
 };
 
 export default memo(App);
-
-// Running these hooks in a separate component ensures we do not inadvertently rerender the entire app when they change.
-const HookIsolator = memo(
-  ({ config, studioInitAction }: { config: PartialAppConfig; studioInitAction?: StudioInitAction }) => {
-    const language = useAppSelector(selectLanguage);
-    const logger = useLogger('system');
-    const dispatch = useAppDispatch();
-
-    // singleton!
-    useReadinessWatcher();
-    useSocketIO();
-    useGlobalModifiersInit();
-    useGlobalHotkeys();
-    useGetOpenAPISchemaQuery();
-    useSyncLoggingConfig();
-
-    useEffect(() => {
-      i18n.changeLanguage(language);
-    }, [language]);
-
-    useEffect(() => {
-      if (size(config)) {
-        logger.info({ config }, 'Received config');
-        dispatch(configChanged(config));
-      }
-    }, [dispatch, config, logger]);
-
-    useEffect(() => {
-      dispatch(appStarted());
-    }, [dispatch]);
-
-    useStudioInitAction(studioInitAction);
-    useStarterModelsToast();
-    useSyncQueueStatus();
-    useFocusRegionWatcher();
-
-    return null;
-  }
-);
-HookIsolator.displayName = 'HookIsolator';
 
 const ModalIsolator = memo(() => {
   return (
